refactor clip_project alg, add optional params
another refactor of the clip project algorithm,
which now supports optional parameters for
user-vendor-project and crs. This means that outputs
will not be reprojected unless an output crs is provided.

The input geometry is now expected to be buffered
and prepared as a WKT area geometry as a preprocessing step,
which can be handled by django, turfjs, or another algorithm